Rename project. Remove unnecessary parameters.
