+++ web-ui/web-v0.8/ctsa-webapp/js/get-suggestions.js
