+++ web-ui/web-v0.8/ctsa-webapp/pom.xml
@@ -8,7 +8,7 @@
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
-	<name>demoWebSpring</name>
+	<name>ctsa-webapp</name>
 	<description>Demo project for Spring Boot</description>
 
 	<parent>
