+++ web-ui/web-v0.8/ctsa-webapp/src/main/resources/static/js/get-suggestions.js
