+++ web-ui/web-v0.8/ctsa-webapp/src/main/resources/templates/summary.html
