+++ web-ui/web-v0.8/ctsa-webapp/src/test/java/io/ctsa/ctsawebapp/DemoWebSpringApplicationTests.java
@@ -1,4 +1,4 @@
-package com.minh.demoWebSpring;
+package io.ctsa.ctsawebapp;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
