--- web-ui/web-v0.8/src/main/resources/templates/modal.html
@@ -1,80 +1,0 @@
-<!DOCTYPE html>
-<html lang="vn">
-
-<head>
-      <meta charset="utf-8" />
-      <meta http-equiv="X-UA-Compatible" content="IE=edge">
-      <title>Dialog</title>
-      <meta name="viewport" content="width=device-width, initial-scale=1">
-
-
-      <!-- Bootstrap -->
-      <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
-      <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
-      <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-
-      <!-- <link rel="stylesheet"
-          href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
-    <link rel="stylesheet"
-          href="/resources/demos/style.css"> -->
-
-      <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-
-      <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
-
-      <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.18.0/axios.min.js"></script>
-      <link rel="stylesheet" href="../../../../css/style.css">
-      <link rel="stylesheet" href="../../../../css/result.css">
-
-      <link rel="stylesheet" href="../../../../css/sweetalert2.min.css">
-      <script src="../../../../js/sweetalert2.all.min.js"></script>
-      <!-- <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet">  -->
-</head>
-
-<body>
-      <a data-toggle="modal" href="" class="btn btn-primary" onClick="successDialog()">Success Dialog</a>
-      <a data-toggle="modal" href="" class="btn btn-primary" onClick="errorDialog()">Error Dialog</a>
-
-      <script>
-            let timerInterval
-            function successDialog() {
-                  //success dialog
-                  swal({
-                        type: 'success',
-                        title: 'Your work has been saved',
-                        showConfirmButton: false,
-                        timer: 1500,
-                        // onOpen: () => {
-                        //     swal.showLoading()
-                        //     timerInterval = setInterval(() => {
-                        //         swal.getContent().querySelector('strong')
-                        //             .textContent = swal.getTimerLeft()
-                        //     }, 100)
-                        // },
-                        // onClose: () => {
-                        //     clearInterval(timerInterval)
-                        // }
-                  })
-            };
-            function errorDialog() {
-                  swal({
-                        type: 'error',
-                        title: 'Please try again later',
-                        showConfirmButton: false,
-                        timer: 1500,
-                        // onOpen: () => {
-                        //     swal.showLoading()
-                        //     timerInterval = setInterval(() => {
-                        //         swal.getContent().querySelector('strong')
-                        //             .textContent = swal.getTimerLeft()
-                        //     }, 100)
-                        // },
-                        // onClose: () => {
-                        //     clearInterval(timerInterval)
-                        // }
-                  })
-            };
-      </script>
-</body>
-
-</html>
